requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
python-dotenv==1.0.0
openpyxl==3.1.2
aiohttp==3.9.1
qrcode[pil]==7.4.2 